feat: move cart logic to upper component
